requirements.txt
flask>=2.3.0
torch>=2.0.0
torchvision>=0.15.0
facenet-pytorch>=2.5.2
faiss-cpu>=1.7.4
numpy>=1.24.0
Pillow>=10.0.0
opencv-python-headless>=4.8.0
scikit-image>=0.21.0
mediapipe>=0.10.0
